requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
jinja2==3.1.4
ffmpeg-python==0.2.0
pyobjc; platform_system == 'Darwin'
